:recycle: extract Celery/Sentry/Langfuse config in infrastructure folder
